requirements.txt
pillow==10.3.0
psutil==5.9.8
